.rainbow
git-svn-id: svn://ultimatepp.org/upp/trunk@11694 f0d560ea-af0d-0410-9eb7-867de7ffcac7

+++ rainbow/SDL20GL/Cursor.cpp
@@ -50,7 +50,7 @@ void  Ctrl::SetMouseCursor(const Image& image)
 			if(sdl_cursor_surface)
 				SDL_FreeSurface(sdl_cursor_surface);
 			
-			int a = image.GetAuxData();
+			int64 a = image.GetAuxData();
 			if(a)
 				sdl_cursor = SDL_CreateSystemCursor(SDL_SystemCursor(a - 1));
 			else {
